requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.7
pytest==5.0.1
Flask-Admin==1.5.3
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
psycopg2-binary==2.8.3
